RestAPI_Tests/requirements.txt
appdirs==1.4.3
asgiref==3.2.5
certifi==2019.11.28
distlib==0.3.0
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
filelock==3.0.12
pipenv==2018.11.26
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.13
virtualenv-clone==0.5.3
